+++ packages/iov-bns/types/bnsconnection.d.ts
@@ -33,7 +33,7 @@ export declare class BnsConnection implements BcpAtomicSwapConnection {
     postTx(tx: PostableBytes): Promise<PostTxResponse>;
     getTicker(ticker: TokenTicker): Promise<BcpQueryEnvelope<BcpTicker>>;
     getAllTickers(): Promise<BcpQueryEnvelope<BcpTicker>>;
-    getAccount(query: BcpAccountQuery): Promise<BcpQueryEnvelope<BcpAccount>>;
+    getAccount(query: BcpAccountQuery): Promise<BcpAccount | undefined>;
     getNonce(query: BcpAddressQuery | BcpPubkeyQuery): Promise<Nonce>;
     /**
      * All matching swaps that are open (from app state)
@@ -82,7 +82,7 @@ export declare class BnsConnection implements BcpAtomicSwapConnection {
     /**
      * Gets current balance and emits an update every time it changes
      */
-    watchAccount(account: BcpAccountQuery): Stream<BcpAccount | undefined>;
+    watchAccount(query: BcpAccountQuery): Stream<BcpAccount | undefined>;
     getBlockchains(query: BnsBlockchainsQuery): Promise<ReadonlyArray<BnsBlockchainNft>>;
     getUsernames(query: BnsUsernamesQuery): Promise<ReadonlyArray<BnsUsernameNft>>;
     protected query(path: string, data: Uint8Array): Promise<QueryResponse>;